Lots of stuff
It's been a while since I uploaded, so I can't recall everything. Need
to keep on that.

I created the Fighter hero, who is in a pretty good spot. The Infested
hero is also pretty well off now. Implemented a custom unit status
frame for the heroes so it uses the LotV template including the shield
bar while using racial skinning.

I deleted the TODO.md file as I believe using issues will be a more
useful and efficient method to keep track of that sort of thing.

An early version of the hero selection has been implemented. In the
future it will be called after a cinematic.

The Zerg enemy are starting to be renamed to Alien to suit the
predecessors. Buildings need to be applied custom Experience rewards,
so a first pass has started toward that. There's still a lot more unit
pre-placement and behavioral additions to go in the future.