Filled map. Opening cinematic. Victory condition.
The map has been filled with Aliens to fight. Nydus Worms occasionally
spawn random aliens that patrol the map.

The first iteration of the opening cinematic is complete. May update in
the future.

Added a simple victory condition for testing purposes.

The map is ready for testing online, privately. Only need to add
Battle.net Information.