Released Swarmer; Updated Hunter; Bug fixes
The spawn area has been adjusted to accomodate new characters.

The Hunter has been reworked somewhat to make it more consistent, including:
* Disruptor Volley replaces the previous final active bility
* Replaced Shadow Blink with Shadow Transfer
* Replaced the Disruptor Volley passive with Psionic Diffusion

The Jumper is a hero that has not yet begun production, but will be taking on the previous divided identity of the Hunter.

The Spawner is a new character that has a unique leveling path. It is considered a difficult character that has the potential to become one of the strongest damage sources.

Many actor bugs have been fixed. It's probable that there are more bugs, please report them if you find them.

The Alien is a new, upcoming hero that will similarly be unique in its leveling.