Fixed upgrades, added push priority, faster levels
Reverted a change to a trigger that gave the heroes to a player instead of replacing them, as it interfered with behavior ownership and thus upgrades.

Heroes now have push priority to make it easier to avoid pathing out of the healing circle.

Doubled the experience gain to make the early game feel less grindy.